requirements.txt
# Core
pandas>=2.0.0
numpy>=1.25.0
pyyaml>=6.0

# Modeling
transformers>=4.42.0
torch>=2.1.0

# Utilities
scikit-learn>=1.3.0